+++ src/components/form/range/range.test.tsx
@@ -23,10 +23,6 @@ import { requiredProps } from '../../../test/required_props';
 
 import { EuiRange } from './range';
 
-jest.mock('./../../../services/accessibility', () => ({
-  htmlIdGenerator: () => () => 'generated-id',
-}));
-
 const props = {
   value: '8',
 };